Move `getEsefAuthority`
